backend/requirements.txt
# Core Django Framework
Django>=5.2,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Astronomy Libraries
stingray>=2.0.0
lightkurve>=2.4.0
astropy>=5.3.0

# Visualization
bokeh>=3.3.0
holoviews>=1.18.0
hvplot>=0.9.0
matplotlib>=3.7.0

# File Handling
astropy-healpix>=1.0.0
h5py>=3.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Development
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
